update packing tests
