+++ .github/workflows/pytest.yaml
@@ -16,7 +16,7 @@ jobs:
         python-version:
           - "3.11"
           - "3.10"
-          - 3."12"
+          - "3.12"
         os:
           - ubuntu-latest
