+++ tests/test_rdkit2ase.py
@@ -1,9 +1,8 @@
 import ase
 import pytest
 import rdkit
-from ase.build import molecule
 
-from rdkit2ase import ase2rdkit, pack, rdkit2ase, smiles2atoms
+from rdkit2ase import ase2rdkit, rdkit2ase, smiles2atoms
 
 
 @pytest.fixture
@@ -66,37 +65,3 @@ def test_smiles_to_atoms(smiles, formula):
     atoms = smiles2atoms(smiles)
     assert isinstance(atoms, ase.Atoms)
     assert atoms.get_chemical_formula() == formula
-
-
-def test_pack_density():
-    atoms = pack("CCO", density=1000)
-    assert atoms.get_chemical_formula() == "C2H6O"
-
-    atoms = pack([("CCO", 2)], density=1000)
-    assert atoms.get_chemical_formula() == "C4H12O2"
-
-    atoms = pack([("CCO", 2), ("O", 1), ("Cl", 3)], density=1000)
-    assert atoms.get_chemical_formula() == "C4H17Cl3O3"
-
-
-def test_pack_box():
-    atoms = pack("CCO", box_size=[5, 5, 5], pbc=False)
-    assert atoms.get_chemical_formula() == "C2H6O"
-    assert atoms.get_volume() == pytest.approx(125)
-
-    # atoms = pack([("CCO", 1)], box_size=[5, 5, 5], pbc=True, tolerance=0)
-    # assert atoms.get_chemical_formula() == "C2H6O"
-    # assert atoms.get_volume() == pytest.approx(125)
-
-    # atoms = pack([("CCO", 1)], box_size=[5, 5, 5], pbc=True, tolerance=0.5)
-    # assert atoms.get_chemical_formula() == "C2H6O"
-    # assert atoms.get_volume() == pytest.approx(125)
-
-
-def test_pack_atoms():
-    atoms = pack([(molecule("CH4"), 2)], density=800)
-    assert atoms.get_chemical_formula() == "C2H8"
-    assert atoms.get_volume() == pytest.approx(66.6, abs=0.001)
-
-
-# TODO assert min distance in pbc!
